Moving trig into numbers subdirectory.
