+++ numbers/trig/trig.go
